removed CaretSaver, now we have save function
