+++ packages/caret/src/index.ts
@@ -1,15 +1,17 @@
-import { CaretSaver } from './CaretSaver';
 import { checkContenteditableSliceForEmptiness } from './checkContenteditableSliceForEmptiness';
 import { extractContenteditableSlice } from './extractContenteditableSlice';
 import { focus } from './focus';
 import { getCaretNodeAndOffset } from './getCaretNodeAndOffset';
+import { getRange } from './getRange';
 import { isCaretAtEndOfInput } from './isCaretAtEndOfInput';
 import { isCaretAtStartOfInput } from './isCaretAtStartOfInput';
+import { save } from './save';
 
-export { CaretSaver,
-  checkContenteditableSliceForEmptiness,
+export { checkContenteditableSliceForEmptiness,
   extractContenteditableSlice,
   focus,
   getCaretNodeAndOffset,
+  getRange,
   isCaretAtEndOfInput,
-  isCaretAtStartOfInput };
+  isCaretAtStartOfInput,
+  save };
